+++ src/integration_tests/regression_test.py
@@ -3,48 +3,8 @@
 from __future__ import annotations
 
 import math
-from collections.abc import Callable
-from typing import Any
 
-from latexify import frontend
-
-
-def _check_function(
-    fn: Callable[..., Any],
-    latex: str,
-    **kwargs,
-) -> None:
-    """Helper to check if the obtained function has the expected LaTeX form.
-
-    Args:
-        fn: Function to check.
-        latex: LaTeX form of `fn`.
-        **kwargs: Arguments passed to `frontend.function`.
-    """
-    # Checks the syntax:
-    #     @function
-    #     def fn(...):
-    #         ...
-    if not kwargs:
-        latexified = frontend.function(fn)
-        assert str(latexified) == latex
-        assert latexified._repr_latex_() == rf"$$ \displaystyle {latex} $$"
-
-    # Checks the syntax:
-    #     @function(**kwargs)
-    #     def fn(...):
-    #         ...
-    latexified = frontend.function(**kwargs)(fn)
-    assert str(latexified) == latex
-    assert latexified._repr_latex_() == rf"$$ \displaystyle {latex} $$"
-
-    # Checks the syntax:
-    #     def fn(...):
-    #         ...
-    #     latexified = function(fn, **kwargs)
-    latexified = frontend.function(fn, **kwargs)
-    assert str(latexified) == latex
-    assert latexified._repr_latex_() == rf"$$ \displaystyle {latex} $$"
+from integration_tests import utils
 
 
 def test_quadratic_solution() -> None:
@@ -52,7 +12,7 @@ def test_quadratic_solution() -> None:
         return (-b + math.sqrt(b**2 - 4 * a * c)) / (2 * a)
 
     latex = r"\mathrm{solve}(a, b, c) = \frac{-b + \sqrt{b^{{2}} - {4} a c}}{{2} a}"
-    _check_function(solve, latex)
+    utils.check_function(solve, latex)
 
 
 def test_sinc() -> None:
@@ -69,7 +29,7 @@ def test_sinc() -> None:
         r"{x = {0}} \\ \frac{\sin{\left({x}\right)}}{x}, & \mathrm{otherwise} "
         r"\end{array} \right."
     )
-    _check_function(sinc, latex)
+    utils.check_function(sinc, latex)
 
 
 def test_x_times_beta() -> None:
@@ -77,11 +37,11 @@ def test_x_times_beta() -> None:
         return x * beta
 
     latex_without_symbols = r"\mathrm{xtimesbeta}(x, beta) = x beta"
-    _check_function(xtimesbeta, latex_without_symbols)
-    _check_function(xtimesbeta, latex_without_symbols, use_math_symbols=False)
+    utils.check_function(xtimesbeta, latex_without_symbols)
+    utils.check_function(xtimesbeta, latex_without_symbols, use_math_symbols=False)
 
     latex_with_symbols = r"\mathrm{xtimesbeta}(x, {\beta}) = x {\beta}"
-    _check_function(xtimesbeta, latex_with_symbols, use_math_symbols=True)
+    utils.check_function(xtimesbeta, latex_with_symbols, use_math_symbols=True)
 
 
 def test_sum_with_limit_1arg() -> None:
@@ -92,7 +52,7 @@ def test_sum_with_limit_1arg() -> None:
         r"\mathrm{sum_with_limit}(n) = \sum_{i = {0}}^{{n - 1}}"
         r" \mathopen{}\left({i^{{2}}}\mathclose{}\right)"
     )
-    _check_function(sum_with_limit, latex)
+    utils.check_function(sum_with_limit, latex)
 
 
 def test_sum_with_limit_2args() -> None:
@@ -103,7 +63,7 @@ def test_sum_with_limit_2args() -> None:
         r"\mathrm{sum_with_limit}(a, n) = \sum_{i = a}^{{n - 1}} "
         r"\mathopen{}\left({i^{{2}}}\mathclose{}\right)"
     )
-    _check_function(sum_with_limit, latex)
+    utils.check_function(sum_with_limit, latex)
 
 
 def test_sum_with_reducible_limit() -> None:
@@ -114,7 +74,7 @@ def test_sum_with_reducible_limit() -> None:
         r"\mathrm{sum_with_limit}(n) = \sum_{i = {0}}^{{n}} "
         r"\mathopen{}\left({i}\mathclose{}\right)"
     )
-    _check_function(sum_with_limit, latex)
+    utils.check_function(sum_with_limit, latex)
 
 
 def test_sum_with_irreducible_limit() -> None:
@@ -125,7 +85,7 @@ def test_sum_with_irreducible_limit() -> None:
         r"\mathrm{sum_with_limit}(n) = \sum_{i = {0}}^{{n {3} - 1}} "
         r"\mathopen{}\left({i}\mathclose{}\right)"
     )
-    _check_function(sum_with_limit, latex)
+    utils.check_function(sum_with_limit, latex)
 
 
 def test_prod_with_limit_1arg() -> None:
@@ -136,7 +96,7 @@ def test_prod_with_limit_1arg() -> None:
         r"\mathrm{prod_with_limit}(n) = "
         r"\prod_{i = {0}}^{{n - 1}} \mathopen{}\left({i^{{2}}}\mathclose{}\right)"
     )
-    _check_function(prod_with_limit, latex)
+    utils.check_function(prod_with_limit, latex)
 
 
 def test_prod_with_limit_2args() -> None:
@@ -147,7 +107,7 @@ def test_prod_with_limit_2args() -> None:
         r"\mathrm{prod_with_limit}(a, n) = "
         r"\prod_{i = a}^{{n - 1}} \mathopen{}\left({i^{{2}}}\mathclose{}\right)"
     )
-    _check_function(prod_with_limit, latex)
+    utils.check_function(prod_with_limit, latex)
 
 
 def test_prod_with_reducible_limits() -> None:
@@ -158,7 +118,7 @@ def test_prod_with_reducible_limits() -> None:
         r"\mathrm{prod_with_limit}(n) = "
         r"\prod_{i = {0}}^{{n - {2}}} \mathopen{}\left({i}\mathclose{}\right)"
     )
-    _check_function(prod_with_limit, latex)
+    utils.check_function(prod_with_limit, latex)
 
 
 def test_prod_with_irreducible_limit() -> None:
@@ -169,14 +129,14 @@ def test_prod_with_irreducible_limit() -> None:
         r"\mathrm{prod_with_limit}(n) = "
         r"\prod_{i = {0}}^{{n {3} - 1}} \mathopen{}\left({i}\mathclose{}\right)"
     )
-    _check_function(prod_with_limit, latex)
+    utils.check_function(prod_with_limit, latex)
 
 
 def test_nested_function() -> None:
     def nested(x):
         return 3 * x
 
-    _check_function(nested, r"\mathrm{nested}(x) = {3} x")
+    utils.check_function(nested, r"\mathrm{nested}(x) = {3} x")
 
 
 def test_double_nested_function() -> None:
@@ -186,20 +146,20 @@ def test_double_nested_function() -> None:
 
         return inner
 
-    _check_function(nested(3), r"\mathrm{inner}(y) = x y")
+    utils.check_function(nested(3), r"\mathrm{inner}(y) = x y")
 
 
 def test_use_raw_function_name() -> None:
     def foo_bar():
         return 42
 
-    _check_function(foo_bar, r"\mathrm{foo_bar}() = {42}")
-    _check_function(
+    utils.check_function(foo_bar, r"\mathrm{foo_bar}() = {42}")
+    utils.check_function(
         foo_bar,
         r"\mathrm{foo_bar}() = {42}",
         use_raw_function_name=False,
     )
-    _check_function(
+    utils.check_function(
         foo_bar,
         r"\mathrm{foo\_bar}() = {42}",
         use_raw_function_name=True,
@@ -211,11 +171,11 @@ def test_reduce_assignments() -> None:
         a = x + x
         return 3 * a
 
-    _check_function(
+    utils.check_function(
         f,
         r"\begin{array}{l} a = x + x \\ \mathrm{f}(x) = {3} a \end{array}",
     )
-    _check_function(
+    utils.check_function(
         f,
         r"\mathrm{f}(x) = {3} \mathopen{}\left( x + x \mathclose{}\right)",
         reduce_assignments=True,
@@ -236,9 +196,9 @@ def test_reduce_assignments_double() -> None:
         r"\end{array}"
     )
 
-    _check_function(f, latex_without_option)
-    _check_function(f, latex_without_option, reduce_assignments=False)
-    _check_function(
+    utils.check_function(f, latex_without_option)
+    utils.check_function(f, latex_without_option, reduce_assignments=False)
+    utils.check_function(
         f,
         r"\mathrm{f}(x) = {3} \mathopen{}\left( x^{{2}} + x^{{2}} \mathclose{}\right)",
         reduce_assignments=True,
@@ -254,7 +214,7 @@ def test_reduce_assignments_with_if() -> None:
         else:
             return n
 
-    _check_function(
+    utils.check_function(
         sigmoid,
         (
             r"\mathrm{sigmoid}(x) = \left\{ \begin{array}{ll} "
@@ -278,137 +238,7 @@ def test_sub_bracket() -> None:
         r"a + b \mathclose{}\right) - \mathopen{}\left( "
         r"a - b \mathclose{}\right) - a b"
     )
-    _check_function(solve, latex)
-
-
-def test_expand_atan2_function() -> None:
-    def solve(x, y):
-        return math.atan2(y, x)
-
-    latex = r"\mathrm{solve}(x, y) = \arctan{\left({\frac{y}{x}}\right)}"
-    _check_function(solve, latex, expand_functions={"atan2"})
-
-
-def test_expand_atan2_nested_function() -> None:
-    def solve(x, y):
-        return math.atan2(math.exp(y), math.exp(x))
-
-    latex = r"\mathrm{solve}(x, y) = \arctan{\left({\frac{e^{y}}{e^{x}}}\right)}"
-    _check_function(solve, latex, expand_functions={"atan2", "exp"})
-
-
-def test_expand_exp_function() -> None:
-    def solve(x):
-        return math.exp(x)
-
-    latex = r"\mathrm{solve}(x) = e^{x}"
-    _check_function(solve, latex, expand_functions={"exp"})
-
-
-def test_expand_exp_nested_function() -> None:
-    def solve(x):
-        return math.exp(math.exp(x))
-
-    latex = r"\mathrm{solve}(x) = e^{e^{x}}"
-    _check_function(solve, latex, expand_functions={"exp"})
-
-
-def test_expand_exp2_function() -> None:
-    def solve(x):
-        return math.exp2(x)
-
-    latex = r"\mathrm{solve}(x) = {2}^{x}"
-    _check_function(solve, latex, expand_functions={"exp2"})
-
-
-def test_expand_exp2_nested_function() -> None:
-    def solve(x):
-        return math.exp2(math.exp2(x))
-
-    latex = r"\mathrm{solve}(x) = {2}^{{2}^{x}}"
-    _check_function(solve, latex, expand_functions={"exp2"})
-
-
-def test_expand_expm1_function() -> None:
-    def solve(x):
-        return math.expm1(x)
-
-    latex = r"\mathrm{solve}(x) = \exp{\left({x}\right)} - {1}"
-    _check_function(solve, latex, expand_functions={"expm1"})
-
-
-def test_expand_expm1_nested_function() -> None:
-    def solve(x, y, z):
-        return math.expm1(math.pow(y, z))
-
-    latex = r"\mathrm{solve}(x, y, z) = e^{y^{z}} - {1}"
-    _check_function(solve, latex, expand_functions={"expm1", "exp", "pow"})
-
-
-def test_expand_hypot_function_without_attribute_access() -> None:
-    from math import hypot
-
-    def solve(x, y, z):
-        return hypot(x, y, z)
-
-    latex = r"\mathrm{solve}(x, y, z) = \sqrt{x^{{2}} + y^{{2}} + z^{{2}}}"
-    _check_function(solve, latex, expand_functions={"hypot"})
-
-
-def test_expand_hypot_function() -> None:
-    def solve(x, y, z):
-        return math.hypot(x, y, z)
-
-    latex = r"\mathrm{solve}(x, y, z) = \sqrt{x^{{2}} + y^{{2}} + z^{{2}}}"
-    _check_function(solve, latex, expand_functions={"hypot"})
-
-
-def test_expand_hypot_nested_function() -> None:
-    def solve(a, b, x, y):
-        return math.hypot(math.hypot(a, b), x, y)
-
-    latex = (
-        r"\mathrm{solve}(a, b, x, y) = "
-        r"\sqrt{"
-        r"\sqrt{a^{{2}} + b^{{2}}}^{{2}} + "
-        r"x^{{2}} + y^{{2}}}"
-    )
-    _check_function(solve, latex, expand_functions={"hypot"})
-
-
-def test_expand_log1p_function() -> None:
-    def solve(x):
-        return math.log1p(x)
-
-    latex = r"\mathrm{solve}(x) = \log{\left({{1} + x}\right)}"
-    _check_function(solve, latex, expand_functions={"log1p"})
-
-
-def test_expand_log1p_nested_function() -> None:
-    def solve(x):
-        return math.log1p(math.exp(x))
-
-    latex = r"\mathrm{solve}(x) = \log{\left({{1} + e^{x}}\right)}"
-    _check_function(solve, latex, expand_functions={"log1p", "exp"})
-
-
-def test_expand_pow_nested_function() -> None:
-    def solve(w, x, y, z):
-        return math.pow(math.pow(w, x), math.pow(y, z))
-
-    latex = (
-        r"\mathrm{solve}(w, x, y, z) = "
-        r"\mathopen{}\left( w^{x} \mathclose{}\right)^{y^{z}}"
-    )
-    _check_function(solve, latex, expand_functions={"pow"})
-
-
-def test_expand_pow_function() -> None:
-    def solve(x, y):
-        return math.pow(x, y)
-
-    latex = r"\mathrm{solve}(x, y) = x^{y}"
-    _check_function(solve, latex, expand_functions={"pow"})
+    utils.check_function(solve, latex)
 
 
 def test_docstring_allowed() -> None:
@@ -417,7 +247,7 @@ def test_docstring_allowed() -> None:
         return x
 
     latex = r"\mathrm{solve}(x) = x"
-    _check_function(solve, latex)
+    utils.check_function(solve, latex)
 
 
 def test_multiple_constants_allowed() -> None:
@@ -428,4 +258,4 @@ def test_multiple_constants_allowed() -> None:
         return x
 
     latex = r"\mathrm{solve}(x) = x"
-    _check_function(solve, latex)
+    utils.check_function(solve, latex)
