+++ src/latexify/config.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import dataclasses
-
 from typing import Any
